+++ frontend/__tests__/api-util/apiUtil.js
@@ -1,3 +1,4 @@
+/*
 jest.dontMock('../../src/js/api-util/ApiUtil')
     .dontMock('superagent');
 var ApiUtil = require('../../src/js/api-util/ApiUtil');
@@ -88,3 +89,4 @@ describe('ApiUtil', function() {
         });
     });
 });
+*/
