+++ frontend/src/js/routes.js
@@ -11,15 +11,15 @@ var routes = (
     <Router history={createBrowserHistory()}>
         <Route path="/" component={require('./components/App')}>
             <IndexRoute component={require('./components/dashboard/Dashboard')}/>
-            <Route path="budgettypes" component={require('./components/budget-type/BudgetMain')}>
-                <IndexRoute component={require('./components/budget-type/BudgetList')}/>
-                <Route path="add" component={require('./components/budget-type/BudgetForm')}/>
-                <Route path="edit/:budgetTypeId" component={require('./components/budget-type/BudgetForm')}/>
+            <Route path="budgettypes" component={require('./components/budget-type/BudgetTypeMain')}>
+                <IndexRoute component={require('./components/budget-type/BudgetTypeList')}/>
+                <Route path="add" component={require('./components/budget-type/BudgetTypeForm')}/>
+                <Route path="edit/:budgetTypeId" component={require('./components/budget-type/BudgetTypeForm')}/>
             </Route>
-            <Route path="projectstatus" component={require('./components/project-status/Main')}>
-                <IndexRoute component={require('./components/project-status/List')}/>
-                <Route path="create" component={require('./components/project-status/Form')}/>
-                <Route path="edit/:id" component={require('./components/project-status/Form')}/>
+            <Route path="projectstatus" component={require('./components/project-status/ProjectStatusMain')}>
+                <IndexRoute component={require('./components/project-status/ProjectStatusList')}/>
+                <Route path="create" component={require('./components/project-status/ProjectStatusForm')}/>
+                <Route path="edit/:id" component={require('./components/project-status/ProjectStatusForm')}/>
             </Route>
         </Route>
     </Router>
